reefatorando: voltando método desviar para personagem
